refactor: extract lastFmSeeder from queueManipulation (phase 1.1)
Extract collectLastFmCandidates function to dedicated lastFmSeeder.ts module as part of Phase 1.1 of queueManipulation refactor. Maintains strict behavior preservation with all 131 existing tests passing.

## Changes
- **New file**: packages/bot/src/utils/music/autoplay/lastFmSeeder.ts (267 LOC)
  - collectLastFmCandidates() - main Last.fm seeding function
  - Internal helper functions for candidate filtering and scoring
  - Dependencies: searchLastFmQuery, calculateRecommendationScore imported from queueManipulation

- **Modified file**: packages/bot/src/utils/music/queueManipulation.ts (-134 LOC)
  - Removed collectLastFmCandidates function body
  - Exported searchLastFmQuery and calculateRecommendationScore for lastFmSeeder dependency
  - Re-imported collectLastFmCandidates for backward compatibility

## Testing
- All 131 queueManipulation tests pass
- No behavior changes - pure extraction
- Circular dependency avoided by keeping searchLastFmQuery and calculateRecommendationScore in queueManipulation

## Phase 1 Status
Part of Lucky refactor plan Phase 1: Breaking up queueManipulation.ts god object.
Prior phases: #658 candidateScorer, #659 diversitySelector, #662 spotifyRecommender
